Se lleva la configuración de las bases de datos de entrada y salida al módulo infrastructure
