+++ application/src/main/java/com/example/batch/application/configuration/DataSourceConfiguration.java
@@ -11,8 +11,6 @@ import org.springframework.transaction.PlatformTransactionManager;
 
 import javax.sql.DataSource;
 
-import static com.example.batch.application.configuration.Constants.ENTRADA_DATASOURCE;
-import static com.example.batch.application.configuration.Constants.SALIDA_DATASOURCE;
 import static com.example.batch.application.configuration.Constants.SPRING_DATASOURCE;
 
 @Configuration
@@ -23,18 +21,6 @@ public class DataSourceConfiguration {
     @BatchDataSource
     @ConfigurationProperties(prefix = SPRING_DATASOURCE)
     public DataSource batchDataSource() {
-        return DataSourceBuilder.create().build();
-    }
-
-    @Bean
-    @ConfigurationProperties(prefix = ENTRADA_DATASOURCE)
-    public DataSource entradaDataSource() {
-        return DataSourceBuilder.create().build();
-    }
-
-    @Bean
-    @ConfigurationProperties(prefix = SALIDA_DATASOURCE)
-    public DataSource salidaDataSource() {
         return DataSourceBuilder.create().build();
     }
