+++ infrastructure/src/main/java/com/example/batch/infrastructure/configuration/Constants.java
@@ -1,0 +1,42 @@
+package com.example.batch.infrastructure.configuration;
+
+public final class Constants {
+
+    private Constants() { }
+
+    public static final String HIBERNATE_HBM_2_DDL_AUTO = "hibernate.hbm2ddl.auto";
+
+    public static final String UPDATE = "update";
+
+    public static final String HIBERNATE_DIALECT = "hibernate.dialect";
+
+    public static final String POSTGRES_SQL_DIALECT = "org.hibernate.dialect.PostgreSQLDialect";
+
+    public static final String ENTRADA = "entrada";
+
+    public static final String ENTRADA_DATASOURCE = "entrada.datasource";
+
+    public static final String ENTRADA_DATA_SOURCE = "entradaDataSource";
+
+    public static final String ENTRADA_ENTITY_MANAGER_FACTORY = "entradaEntityManagerFactory";
+
+    public static final String ENTRADA_TRANSACTION_MANAGER = "entradaTransactionManager";
+
+    public static final String ENTRADA_ENTITY_PACKAGE = "com.example.batch.infrastructure.entity.entrada";
+
+    public static final String ENTRADA_REPOSITORY_PACKAGE = "com.example.batch.infrastructure.repository.entrada";
+
+    public static final String SALIDA = "salida";
+
+    public static final String SALIDA_DATASOURCE = "salida.datasource";
+
+    public static final String SALIDA_DATA_SOURCE = "salidaDataSource";
+
+    public static final String SALIDA_ENTITY_MANAGER_FACTORY = "salidaEntityManagerFactory";
+
+    public static final String SALIDA_TRANSACTION_MANAGER = "salidaTransactionManager";
+
+    public static final String SALIDA_ENTITY_PACKAGE = "com.example.batch.infrastructure.entity.salida";
+
+    public static final String SALIDA_REPOSITORY_PACKAGE = "com.example.batch.infrastructure.repository.salida";
+}
