+++ infrastructure/src/main/java/com/example/batch/infrastructure/configuration/SalidaJpaConfiguration.java
@@ -1,6 +1,9 @@
-package com.example.batch.application.configuration;
+package com.example.batch.infrastructure.configuration;
 
+import jakarta.persistence.EntityManagerFactory;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.jdbc.DataSourceBuilder;
 import org.springframework.boot.orm.jpa.EntityManagerFactoryBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -10,19 +13,19 @@ import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 import org.springframework.transaction.PlatformTransactionManager;
 
 import javax.sql.DataSource;
-import jakarta.persistence.EntityManagerFactory;
 import java.util.Map;
 
-import static com.example.batch.application.configuration.Constants.HIBERNATE_DIALECT;
-import static com.example.batch.application.configuration.Constants.HIBERNATE_HBM_2_DDL_AUTO;
-import static com.example.batch.application.configuration.Constants.POSTGRES_SQL_DIALECT;
-import static com.example.batch.application.configuration.Constants.SALIDA;
-import static com.example.batch.application.configuration.Constants.SALIDA_DATA_SOURCE;
-import static com.example.batch.application.configuration.Constants.SALIDA_ENTITY_MANAGER_FACTORY;
-import static com.example.batch.application.configuration.Constants.SALIDA_ENTITY_PACKAGE;
-import static com.example.batch.application.configuration.Constants.SALIDA_REPOSITORY_PACKAGE;
-import static com.example.batch.application.configuration.Constants.SALIDA_TRANSACTION_MANAGER;
-import static com.example.batch.application.configuration.Constants.UPDATE;
+import static com.example.batch.infrastructure.configuration.Constants.HIBERNATE_DIALECT;
+import static com.example.batch.infrastructure.configuration.Constants.HIBERNATE_HBM_2_DDL_AUTO;
+import static com.example.batch.infrastructure.configuration.Constants.POSTGRES_SQL_DIALECT;
+import static com.example.batch.infrastructure.configuration.Constants.SALIDA;
+import static com.example.batch.infrastructure.configuration.Constants.SALIDA_DATASOURCE;
+import static com.example.batch.infrastructure.configuration.Constants.SALIDA_DATA_SOURCE;
+import static com.example.batch.infrastructure.configuration.Constants.SALIDA_ENTITY_MANAGER_FACTORY;
+import static com.example.batch.infrastructure.configuration.Constants.SALIDA_ENTITY_PACKAGE;
+import static com.example.batch.infrastructure.configuration.Constants.SALIDA_REPOSITORY_PACKAGE;
+import static com.example.batch.infrastructure.configuration.Constants.SALIDA_TRANSACTION_MANAGER;
+import static com.example.batch.infrastructure.configuration.Constants.UPDATE;
 
 @Configuration
 @EnableJpaRepositories(
@@ -31,6 +34,12 @@ import static com.example.batch.application.configuration.Constants.UPDATE;
         transactionManagerRef = SALIDA_TRANSACTION_MANAGER
 )
 public class SalidaJpaConfiguration {
+
+    @Bean
+    @ConfigurationProperties(prefix = SALIDA_DATASOURCE)
+    public DataSource salidaDataSource() {
+        return DataSourceBuilder.create().build();
+    }
 
     @Bean
     public LocalContainerEntityManagerFactoryBean salidaEntityManagerFactory(
