+++ infrastructure/src/test/java/com/example/batch/infrastructure/configuration/EntradaJpaConfigurationTest.java
@@ -1,4 +1,4 @@
-package com.example.batch.application.configuration;
+package com.example.batch.infrastructure.configuration;
 
 import jakarta.persistence.EntityManagerFactory;
 import org.junit.jupiter.api.Test;
@@ -7,29 +7,39 @@ import org.springframework.boot.orm.jpa.EntityManagerFactoryBuilder;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 
 import javax.sql.DataSource;
-
 import java.util.Map;
 
-import static com.example.batch.application.configuration.Constants.HIBERNATE_DIALECT;
-import static com.example.batch.application.configuration.Constants.HIBERNATE_HBM_2_DDL_AUTO;
-import static com.example.batch.application.configuration.Constants.POSTGRES_SQL_DIALECT;
-import static com.example.batch.application.configuration.Constants.SALIDA;
-import static com.example.batch.application.configuration.Constants.SALIDA_ENTITY_PACKAGE;
-import static com.example.batch.application.configuration.Constants.UPDATE;
-import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertSame;
+import static com.example.batch.infrastructure.configuration.Constants.ENTRADA;
+import static com.example.batch.infrastructure.configuration.Constants.ENTRADA_ENTITY_PACKAGE;
+import static com.example.batch.infrastructure.configuration.Constants.HIBERNATE_DIALECT;
+import static com.example.batch.infrastructure.configuration.Constants.HIBERNATE_HBM_2_DDL_AUTO;
+import static com.example.batch.infrastructure.configuration.Constants.POSTGRES_SQL_DIALECT;
+import static com.example.batch.infrastructure.configuration.Constants.UPDATE;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 
-class SalidaJpaConfigurationTest {
+class EntradaJpaConfigurationTest {
 
-    private final SalidaJpaConfiguration salidaJpaConfiguration = new SalidaJpaConfiguration();
+    private final EntradaJpaConfiguration entradaJpaConfiguration = new EntradaJpaConfiguration();
 
     @Test
-    void salidaEntityManagerFactory() {
+    void entradaDataSource() {
+        // when
+        var actual = entradaJpaConfiguration.entradaDataSource();
+
+        // then
+        assertAll(
+                () -> assertNotNull(actual),
+                () -> assertInstanceOf(DataSource.class, actual)
+        );
+    }
+
+    @Test
+    void entradaEntityManagerFactory() {
         // given
         var entityManagerFactoryBuilder = Mockito.mock(EntityManagerFactoryBuilder.class);
         var builder = Mockito.mock(EntityManagerFactoryBuilder.Builder.class);
@@ -38,13 +48,13 @@ class SalidaJpaConfigurationTest {
         var expected = Mockito.mock(LocalContainerEntityManagerFactoryBean.class);
 
         when(entityManagerFactoryBuilder.dataSource(dataSource)).thenReturn(builder);
-        when(builder.packages(SALIDA_ENTITY_PACKAGE)).thenReturn(builder);
-        when(builder.persistenceUnit(SALIDA)).thenReturn(builder);
+        when(builder.packages(ENTRADA_ENTITY_PACKAGE)).thenReturn(builder);
+        when(builder.persistenceUnit(ENTRADA)).thenReturn(builder);
         when(builder.properties(properties)).thenReturn(builder);
         when(builder.build()).thenReturn(expected);
 
         // when
-        var actual = salidaJpaConfiguration.salidaEntityManagerFactory(entityManagerFactoryBuilder, dataSource);
+        var actual = entradaJpaConfiguration.entradaEntityManagerFactory(entityManagerFactoryBuilder, dataSource);
 
         // then
         assertAll(
@@ -53,20 +63,20 @@ class SalidaJpaConfigurationTest {
         );
 
         verify(entityManagerFactoryBuilder, times(1)).dataSource(dataSource);
-        verify(builder, times(1)).packages(SALIDA_ENTITY_PACKAGE);
-        verify(builder, times(1)).persistenceUnit(SALIDA);
+        verify(builder, times(1)).packages(ENTRADA_ENTITY_PACKAGE);
+        verify(builder, times(1)).persistenceUnit(ENTRADA);
         verify(builder, times(1)).properties(properties);
         verify(builder, times(1)).build();
         verifyNoMoreInteractions(entityManagerFactoryBuilder, builder);
     }
 
     @Test
-    void salidaTransactionManager() {
+    void entradaTransactionManager() {
         // given
         var entityManagerFactory = Mockito.mock(EntityManagerFactory.class);
 
         // when
-        var actual = salidaJpaConfiguration.salidaTransactionManager(entityManagerFactory);
+        var actual = entradaJpaConfiguration.entradaTransactionManager(entityManagerFactory);
 
         // then
         assertNotNull(actual);
